requirements.txt
fastapi
uvicorn[standard]
PyGithub
python-dotenv